DOC: Improve docstrings and comments in secondary.py
Functions have become slightly more complex due to the addition of
partial verification client support in the reference implementation,
so the documentation had to be improved.

Note that some of the procedures and variables described are slightly
different than what the docstring suggests: the changes to functions
in the next commit will bring them in line.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>