Update metadata distributed to partial verification secondaries
Now metadata distributed to the partial verification secondaries
would be in the form of a archive with the structure almost
similar to that of the distributed archive for full verification.
Here only director repo would be present and it would only have
the targets metadata.

This is done so that the secondary could obtain, save and expand
the metadata recieved from primary in a similar fashion as that
of full verification. It would further help in simplifying the
partial verification process